+++ packages/nimble-components/src/table/components/group-row/styles.ts
@@ -14,15 +14,32 @@ import {
 import { Theme } from '../../../theme-provider/types';
 import { hexToRgbaCssColor } from '../../../utilities/style/colors';
 import { themeBehavior } from '../../../utilities/style/theme';
+import { userSelectNone } from '../../../utilities/style/user-select';
 
 export const styles = css`
-    ${display('flex')}
+    ${display('grid')}
 
     :host {
         align-items: center;
         height: calc(${controlHeight} + 2 * ${borderWidth});
         border-top: calc(2 * ${borderWidth}) solid ${applicationBackgroundColor};
         box-sizing: border-box;
+        grid-template-columns:
+            calc(
+                ${controlHeight} *
+                    (var(--ni-private-table-group-row-indent-level) + 1)
+            )
+            1fr;
+    }
+
+    :host([selectable]) {
+        grid-template-columns:
+            ${controlHeight}
+            calc(
+                ${controlHeight} *
+                    (var(--ni-private-table-group-row-indent-level) + 1)
+            )
+            1fr;
     }
 
     :host([expanded]) .animating,
@@ -48,7 +65,6 @@ export const styles = css`
             ${smallPadding} * 2 + ${standardPadding} * 2 *
                 var(--ni-private-table-group-row-indent-level)
         );
-        width: ${controlSlimHeight};
         height: ${controlSlimHeight};
     }
 
@@ -67,15 +83,16 @@ export const styles = css`
 
     .group-header-view {
         padding-left: calc(${standardPadding} / 2);
-        user-select: none;
+        ${userSelectNone}
         overflow: hidden;
         display: flex;
     }
 
     .group-row-child-count {
         padding-left: 2px;
+        padding-right: calc(${standardPadding} / 2);
         pointer-events: none;
-        user-select: none;
+        ${userSelectNone}
     }
 
     @media (prefers-reduced-motion) {