+++ packages/nimble-blazor/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/nimble-blazor
 
-This log was last generated on Fri, 28 Apr 2023 15:33:54 GMT and should not be manually modified.
+This log was last generated on Mon, 15 May 2023 14:59:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 11.8.29
+
+Mon, 15 May 2023 14:59:04 GMT
+
+### Patches
+
+- Ensure components can take additional attributes ([ni/nimble@c582727](https://github.com/ni/nimble/commit/c5827276a724ddd368d0acf35b1d29956971639f))
+
+## 11.8.20
+
+Tue, 02 May 2023 15:07:15 GMT
+
+### Patches
+
+- Create Blazor Hybrid example app. Provide workaround for Blazor Hybrid to work with Nimble Blazor. ([ni/nimble@da1d0f2](https://github.com/ni/nimble/commit/da1d0f27d9b555f2b7df2d112122dda0520826f0))
 
 ## 11.8.18